test(postgresql): e2e for TLS options
Adds e2e tests for the functionality of providing the TLS options by
testing with TLS turned on on the server side using the
`tls.enabled=true` and `tls.autoGenerated=true` PostgreSQL Chart
parameters.

To make the tests as robust as possible: port forwarding is not used,
but rather exposing the server port via NodePort; timeouts are increased
on kubectl wait's, from 2 min to 3 min.

The test iself reuses parts of the existing tests, and adds one
additional check that asserts if TLS is used.